Yolo and binary image implementation

+++ YOLO.py
@@ -6,10 +6,11 @@ from ultralytics import YOLO
 import json
 class ObjectDetection:
 
-    def __init__(self, capture_index, webcam=True, image_path=None):
+    def __init__(self, capture_index, webcam=True, image_path=None, debug_mode=False):
         self.capture_index = capture_index
         self.webcam = webcam
         self.image_path = image_path
+        self.debug_mode = debug_mode
 
         # ToDo: Make everything run on cuda
         self.device = 'cuda' if torch.cuda.is_available() else 'cpu'
@@ -36,18 +37,11 @@ class ObjectDetection:
             boxes = result.boxes.cpu().numpy()
 
             # Extract xy coordinates, confidences and class ids
-            xyxys.append(boxes.xyxy)
+            xyxys.append(boxes.xyxy.tolist())
             confidences.append(boxes.conf.tolist())  # Convert ndarray to list
             class_ids.append(boxes.cls.tolist())  # Convert ndarray to list
 
-        # Save confidences and class_ids to a JSON file
-        data = {
-            "confidences": confidences,
-            "class_ids": class_ids
-        }
 
-        with open('results.json', 'w') as f:
-            json.dump(data, f)
 
         return results[0].plot(), xyxys, confidences, class_ids
 
@@ -91,7 +85,36 @@ class ObjectDetection:
             # Draw bounding boxes on the frame and get the plotted image, bounding boxes, confidences, and class ids
             plotted_image, xyxys, confidences, class_ids = self.plot_bboxes(results, frame)
 
-            # Display the resulting frame
-            cv2.imshow('frame', plotted_image)
-            cv2.waitKey(0)  # Wait for any key to be pressed to close the window
-            cv2.destroyAllWindows()
+            # Save confidences and class_ids to a JSON file
+            data = {
+                "confidences": confidences,
+                "class_ids": class_ids,
+                "bounding_boxes": xyxys
+            }
+
+            with open('results.json', 'w') as f:
+                json.dump(data, f)
+
+            if self.debug_mode:
+                # Display the resulting frame
+                cv2.imshow('thisframe', plotted_image)
+                cv2.waitKey(0)  # Wait for any key to be pressed to close the window
+                cv2.destroyAllWindows()
+
+    def get_Yolo_Results(self):
+        """
+        Returns the results of the object detection process.
+
+        Returns
+        -------
+        list, list, list
+            the bounding box coordinates, the confidences, and the class ids
+        """
+        # Read bounding box coordinates
+        with open('results.json', 'r') as f:
+            data = json.load(f)
+
+        confidences = data['confidences']
+        class_ids = data['class_ids']
+        bounding_boxes = data['bounding_boxes']
+        return bounding_boxes, confidences, class_ids
